fix: refactor package.json and place at root folder
Since `node_modules` would get quite large for both the client
and the server, it would be a better decision to move the
`package.json` with dependencies to the root folder, in case we happen to
have common packages for both server and client.

For scripts, server and client can have their scripts separately in
their own `package.json`.